cortex_gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
python-dotenv
PyPDF2
python-multipart
